feat: move some pages to the app router
