+++ client/app/report/page.tsx
@@ -1,20 +1,20 @@
+"use client";
+
 import { useFormState } from "react-use-form-state";
 import { Flex } from "rebass/styled-components";
 import React, { useState } from "react";
 import axios from "axios";
 
-import Text, { H2, Span } from "../components/Text";
-import AppWrapper from "../components/AppWrapper";
-import { TextInput } from "../components/Input";
-import { Button } from "../components/Button";
-import { Col } from "../components/Layout";
-import Icon from "../components/Icon";
-import { useMessage } from "../hooks";
-import { APIv2 } from "../consts";
+import Text, { H2, Span } from "../../components/Text";
+import AppWrapper from "../../components/AppWrapper";
+import { TextInput } from "../../components/Input";
+import { Button } from "../../components/Button";
+import { Col } from "../../components/Layout";
+import Icon from "../../components/Icon";
+import { useMessage } from "../../hooks";
+import { APIv2 } from "../../consts";
 
-import getConfig from "next/config";
-
-const { publicRuntimeConfig } = getConfig();
+import { publicRuntimeConfig } from '../../../next.config';
 
 const ReportPage = () => {
   const [formState, { text }] = useFormState<{ url: string }>();
